added support for external resources in graph resolution
External resources are just "markers" to add passes to the dependency
chain of the graph.

If a compute pass just modifies an external buffer but does not create
any resources on its own, it could previously not be part of the
execution chain.

Now with external resources new "names" can be created and moved.
That allows side-effect only passes to be part of the cool kids and
be executed as well!